More refactoring for Mantis support
Refactor handling of signed transactions around SignedTransactionInfo,
a type not from the json rpc interface, that carries from and hash as well as
v r s and the common fields nonce to data gas gasPrice.
This refactoring is possible because we've stopped
using the Ethereum node as a wallet.

Make use of EIP-145 configurable.

Add an entry for the shared Cardano KEVM Devnet.

(Disabled) Add a logback configuration for the 2020-evm image to dump execution details.

Stop using 'pending, it doesn't currently work on Mantis. Use 'latest.

Edit TODO