Be stricter about checking the deposited amount.
Specifically, rather than demanding that the deposited amount is *at
least* what is expected, demand that it is *exactly* what is expected.

The rationale for this is that, when we move to multiplexing
interactions on a single contract:

- The interaction balance will have to be part of the merkelized state,
  since we can't query the VM for it.
- Therefore, participants will have to supply the correct value when
  restoring that state.
- So, if we allow participants to over-deposit, they could potentially
  throw off that value from what is expected, introducing a possible DoS
  vulnerability for the other participant unless we introduce some
  fiddly logic to deal with the possibility. It seems simpler to catch
  this up front.

I have confirmed that the downstream tests for glow still pass with this
change, in addition to the tests in this repo.