Delete define-*-logging and &check-sufficient-deposit0!
These are no longer called by glow, and would need reworking for
correctness: firstly, we were only still defining
&check-sufficient-deposit0!, while other assets need checking too.
Secondly, this assumes the native token, but in practice we will
need to dispatch based on the type of asset.

All that is done inside glow now, so let's just delete these.