ensure-addresses-prefunded don't batch transfers.
This doesn't work for erc20 assets, because the transfers end up
being attempted by the batch contract, rather than the funder,
whose authorization is required.

We could still do this for the native asset, but this is only used
for testing and only saves a small amount of time if its the native
asset, so we leave this for future work.

...Still need to update the tests to not assume the batch parameter.